Update to ExoPlayer 2.4.0k1
Fixes #149
