+++ app/build.gradle
@@ -77,8 +77,9 @@ dependencies {
     compile 'com.android.support:support-v4:25.3.1'
     compile 'com.android.support:leanback-v17:25.3.1'
     compile 'com.android.support:preference-leanback-v17:25.3.1'
-    compile 'com.google.android.exoplayer:exoplayer:r2.3.0k1'
-    compile 'com.google.android.exoplayer:extension-ffmpeg:r2.3.0k1'
+    compile 'com.google.android.exoplayer:exoplayer-core:r2.4.0k1'
+    compile 'com.google.android.exoplayer:exoplayer-ui:r2.4.0k1'
+    compile 'com.google.android.exoplayer:extension-ffmpeg:r2.4.0k1'
     // Used for testing local exoplayer builds
 //    compile(name: 'library-debug', ext: 'aar')
     // Used for testing local exoplayer-ffmpeg builds
